Corrects flaky error occuring in CollectTypesController spec.
