+++ app/controllers/hyrax/admin/collection_types_controller_override.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module Hyrax
+  module Admin
+    module CollectionTypesControllerOverride
+       # [Hyrax-overwrite-v3.0.0.pre.rc1] Adding deposit_only collection to the
+       # permitted params.
+      def collection_type_params
+        params.require(:collection_type).permit(
+          :title, :description, :nestable, :brandable, :discoverable, :sharable,
+          :share_applies_to_new_works, :allow_multiple_membership, :require_membership,
+          :assigns_workflow, :assigns_visibility, :badge_color, :deposit_only_collection
+        )
+      end
+    end
+  end
+end
